Moved NamedPipes into the routing static lib
