+++ src/loader/MacroQuest.vcxproj
@@ -125,7 +125,6 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="..\common\HotKeys.cpp" />
-    <ClCompile Include="..\common\NamedPipes.cpp" />
     <ClCompile Include="AutoLogin.cpp" />
     <ClCompile Include="Crashpad.cpp" />
     <ClCompile Include="HotKeyControl.cpp" />
@@ -140,8 +139,6 @@
     <ClInclude Include="..\..\include\mq\base\BuildInfo.h" />
     <ClInclude Include="..\..\include\mq\utils\Naming.h" />
     <ClInclude Include="..\common\HotKeys.h" />
-    <ClInclude Include="..\common\NamedPipes.h" />
-    <ClInclude Include="..\common\NamedPipesProtocol.h" />
     <ClInclude Include="Crashpad.h" />
     <ClInclude Include="HotKeyControl.h" />
     <ClInclude Include="MacroQuest.h" />
